Tentativa de corrigir erro em banco de dados

+++ src/main/java/com/marcos/primeirocrud/controller/LivroController.java
@@ -38,7 +38,7 @@ public class LivroController {
     //DELETE
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deletaLivro(@PathVariable Long id){
+    public void deletaLivro(@PathVariable Integer id){
         livroService.deletaLivro(id);
     }
