+++ src/main/java/com/marcos/primeirocrud/repository/LivroRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Collection;
 import java.util.List;
 
-public interface LivroRepository extends JpaRepository<Livro, Long> {
+public interface LivroRepository extends JpaRepository<Livro, Integer> {
     List<Livro> findByDescricaoContainingIgnoreCase(String descricao);
 }
